refactor: Providers を shell+composition に分解
